StoryView rename to StoriesView and StoriesView refactoring
